github_analyser--v1/requirements.txt
fastapi
uvicorn
requests
streamlit
beautifulsoup4
pandas
python-dotenv
google-generativeai
langchain-google-genai
langchain
PyPDF2
